requirements.txt
fastapi
uvicorn[standard]

requests
psycopg2-binary
PyYAML
pydantic
